This fixes a race condition (revealed by my other check-in r4067) in the romtool by changing the Makefiles to be no longer recursive (once again, recursive make is to be considered harmful).  Tried to (quickly) unify most of the Makefile code, but medium-term this is going to be worked on for Kconfig support anyways.
Also fix a sign cast error in rom-mkpayload in case people want to compile this
with -W -Werror

Patch relative to coreboot-v2/util/romtool

and
Acked-by: Stefan Reinauer <stepan@coresystems.de>
in order to get the tree working decently asap



git-svn-id: svn://svn.coreboot.org/coreboot/trunk@4069 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ util/romtool/tools/rom-mkpayload.c
@@ -163,7 +163,7 @@ int parse_elf(unsigned char *input, unsigned char **output, int algo,
 		/* If the compressed section is larger, then use the
 		   original stuff */
 
-		if (len > phdr[i].p_filesz) {
+		if ((unsigned int)len > phdr[i].p_filesz) {
 			segs[segments].compression = 0;
 			segs[segments].len = htonl(phdr[i].p_filesz);
